fix: refactore and improve seo
