+++ src/app/components/sub/header/NavBar.jsx
@@ -1,9 +1,8 @@
 'use client';
-
 import { useState } from 'react';
 import MobileNav from './MobileNav';
-import HeaderLogo from '../sub/header/HeaderLogo';
-import BurgerButton from '../sub/header/BurgerButton';
+import HeaderLogo from '../../sub/header/HeaderLogo';
+import BurgerButton from '../../sub/header/BurgerButton';
 import DesktopNav from './DesktopNav';
 
 const NavBar = () => {
@@ -12,11 +11,15 @@ const NavBar = () => {
   const toggleMenu = () => setNavBarIsOpen(!navBarIsOpen);
 
   return (
-    <nav className="w-full px-4">
+    <nav className="w-full px-4" aria-label="Main Navigation">
       <div className="flex justify-between items-center">
         <HeaderLogo />
         <DesktopNav />
-        <BurgerButton toggleMenu={toggleMenu} navBarIsOpen={navBarIsOpen} />
+        <BurgerButton
+          toggleMenu={toggleMenu}
+          navBarIsOpen={navBarIsOpen}
+          aria-label="Toggle mobile menu"
+        />
       </div>
       <MobileNav
         setNavBarIsOpen={setNavBarIsOpen}
